feat: add since flag for import (#57)
Closes #38

PR adds a `--since` flag to the import entrypoint, so that a user can
now easily import a subset of their message data. The flag accepts both
absolute dates (e.g. `2025-09-09`) as well as durations (e.g. `1M` for 1
month). For the duration string, the parser accepts both uppercase and
lowercase, and allows for `D` (days), `W` (weeks), `M` (months), and `Y`
(years) and attempts to make the conversion "calendar-aware", such that
if you do `1M` at the end of a month, then it will set the date to the
last day of the preceding month, accounting for different number of
days.

Signed-off-by: Matthew Peveler <mpeveler@tigerdata.com>